Move monitoring in the fs to its own package
